+++ core/src/db/insert/VectorSource.h
@@ -29,14 +29,17 @@ namespace engine {
 
 class VectorSource {
  public:
-    VectorSource(const size_t& n, const float* vectors);
+    explicit VectorSource(VectorsData& vectors);
 
     Status
     Add(const ExecutionEnginePtr& execution_engine, const meta::TableFileSchema& table_file_schema,
-        const size_t& num_vectors_to_add, size_t& num_vectors_added, IDNumbers& vector_ids);
+        const size_t& num_vectors_to_add, size_t& num_vectors_added);
 
     size_t
     GetNumVectorsAdded();
+
+    size_t
+    SingleVectorSize(uint16_t dimension);
 
     bool
     AllAdded();
@@ -45,8 +48,7 @@ class VectorSource {
     GetVectorIds();
 
  private:
-    const size_t n_;
-    const float* vectors_;
+    VectorsData& vectors_;
     IDNumbers vector_ids_;
 
     size_t current_num_vectors_added;