+++ CHANGELOG.md
@@ -16,12 +16,13 @@ Please mark all change in change log and use the issue from GitHub
 - \#216 - Add CLI to get server info
 - \#343 - Add Opentracing
 - \#665 - Support get/set config via CLI
+- \#759 - Put C++ sdk out of milvus/core
 - \#766 - If partition tag is similar, wrong partition is searched
 - \#771 - Add server build commit info interface
-- \#759 - Put C++ sdk out of milvus/core
 - \#788 - Add web server into server module
 - \#813 - Add push mode for prometheus monitor
 - \#815 - Support MinIO storage
+- \#823 - Support binary vector tanimoto/jaccard/hamming metric
 - \#910 - Change Milvus c++ standard to c++17
 
 ## Improvement